+++ .eslintrc.js
@@ -11,6 +11,7 @@ module.exports = {
     '@typescript-eslint/consistent-type-definitions': 'error',
     '@typescript-eslint/member-ordering': 'error',
     '@typescript-eslint/no-inferrable-types': 'error',
+    '@typescript-eslint/no-redeclare': 'error',
     '@typescript-eslint/no-unnecessary-type-assertion': 'error',
     '@typescript-eslint/no-unused-vars': [
       'error',
@@ -32,7 +33,9 @@ module.exports = {
     ],
     'import/no-namespace': 'off',
     'import/no-named-export': 'off',
+    'no-redeclare': 'off',
     'no-unused-vars': 'off',
+    'prefer-object-spread': 'off',
     'spaced-comment': 'off',
     'prettier/prettier': [
       'error',
@@ -44,5 +47,13 @@ module.exports = {
         plugins: ['prettier-plugin-package']
       }
     ]
-  }
+  },
+  overrides: [
+    {
+      "files": ["**/*.ts"],
+      "rules": {
+        "no-undef": "off"
+      }
+    }
+  ]
 };