+++ packages/commonjs/src/transform-commonjs.js
@@ -11,6 +11,7 @@ import {
   isDefineCompiledEsm,
   isFalsy,
   isReference,
+  isShorthandProperty,
   isTruthy,
   KEY_COMPILED_ESM
 } from './ast-utils';
@@ -25,9 +26,14 @@ import {
   isRequireStatement,
   isStaticRequireStatement
 } from './generate-imports';
-import { DYNAMIC_JSON_PREFIX, DYNAMIC_REGISTER_PREFIX } from './helpers';
+import { DYNAMIC_JSON_PREFIX } from './helpers';
 import { tryParse } from './parse';
 import { deconflict, getName, getVirtualPathForDynamicRequirePath } from './utils';
+import {
+  isModuleRegisterProxy,
+  unwrapModuleRegisterProxy,
+  wrapModuleRegisterProxy
+} from './dynamic-packages-manager';
 
 const exportsPattern = /^(?:module\.)?exports(?:\.([a-zA-Z_$][a-zA-Z_$0-9]*))?$/;
 
@@ -40,6 +46,7 @@ export default function transformCommonjs(
   isEsModule,
   ignoreGlobal,
   ignoreRequire,
+  getIgnoreTryCatchRequireStatementMode,
   sourceMap,
   isDynamicRequireModulesEnabled,
   dynamicRequireModuleSet,
@@ -61,6 +68,7 @@ export default function transformCommonjs(
   let scope = attachScopes(ast, 'scope');
   let lexicalDepth = 0;
   let programDepth = 0;
+  let currentTryBlockEnd = null;
   let shouldWrap = false;
   const defineCompiledEsmExpressions = [];
 
@@ -70,6 +78,7 @@ export default function transformCommonjs(
   const HELPERS_NAME = deconflict(scope, globals, 'commonjsHelpers');
   const namedExports = {};
   const dynamicRegisterSources = new Set();
+  let hasRemovedRequire = false;
 
   const {
     addRequireStatement,
@@ -94,6 +103,10 @@ export default function transformCommonjs(
         return;
       }
 
+      if (currentTryBlockEnd !== null && node.start > currentTryBlockEnd) {
+        currentTryBlockEnd = null;
+      }
+
       programDepth += 1;
       if (node.scope) ({ scope } = node);
       if (functionType.test(node.type)) lexicalDepth += 1;
@@ -104,6 +117,11 @@ export default function transformCommonjs(
 
       // eslint-disable-next-line default-case
       switch (node.type) {
+        case 'TryStatement':
+          if (currentTryBlockEnd === null) {
+            currentTryBlockEnd = node.block.end;
+          }
+          return;
         case 'AssignmentExpression':
           if (node.left.type === 'MemberExpression') {
             const flattened = getKeypath(node.left);
@@ -157,6 +175,32 @@ export default function transformCommonjs(
             }
             return;
           }
+
+          if (
+            node.callee.object &&
+            node.callee.object.name === 'require' &&
+            node.callee.property.name === 'resolve' &&
+            hasDynamicModuleForPath(id, '/', dynamicRequireModuleSet)
+          ) {
+            const requireNode = node.callee.object;
+            magicString.appendLeft(
+              node.end - 1,
+              `,${JSON.stringify(
+                dirname(id) === '.' ? null /* default behavior */ : virtualDynamicRequirePath
+              )}`
+            );
+            magicString.overwrite(
+              requireNode.start,
+              requireNode.end,
+              `${HELPERS_NAME}.commonjsRequire`,
+              {
+                storeName: true
+              }
+            );
+            uses.commonjsHelpers = true;
+            return;
+          }
+
           if (!isStaticRequireStatement(node, scope)) return;
           if (!isDynamicRequireModulesEnabled) {
             skippedNodes.add(node.callee);
@@ -165,35 +209,61 @@ export default function transformCommonjs(
             skippedNodes.add(node.callee);
             const usesReturnValue = parent.type !== 'ExpressionStatement';
 
+            let canConvertRequire = true;
+            let shouldRemoveRequireStatement = false;
+
+            if (currentTryBlockEnd !== null) {
+              ({
+                canConvertRequire,
+                shouldRemoveRequireStatement
+              } = getIgnoreTryCatchRequireStatementMode(node.arguments[0].value));
+
+              if (shouldRemoveRequireStatement) {
+                hasRemovedRequire = true;
+              }
+            }
+
             let sourceId = getRequireStringArg(node);
-            const isDynamicRegister = sourceId.startsWith(DYNAMIC_REGISTER_PREFIX);
+            const isDynamicRegister = isModuleRegisterProxy(sourceId);
             if (isDynamicRegister) {
-              sourceId = sourceId.substr(DYNAMIC_REGISTER_PREFIX.length);
+              sourceId = unwrapModuleRegisterProxy(sourceId);
               if (sourceId.endsWith('.json')) {
                 sourceId = DYNAMIC_JSON_PREFIX + sourceId;
               }
-              dynamicRegisterSources.add(sourceId);
+              dynamicRegisterSources.add(wrapModuleRegisterProxy(sourceId));
             } else {
               if (
                 !sourceId.endsWith('.json') &&
                 hasDynamicModuleForPath(sourceId, id, dynamicRequireModuleSet)
               ) {
-                magicString.overwrite(
-                  node.start,
-                  node.end,
-                  `${HELPERS_NAME}.commonjsRequire(${JSON.stringify(
-                    getVirtualPathForDynamicRequirePath(sourceId, commonDir)
-                  )}, ${JSON.stringify(
-                    dirname(id) === '.' ? null /* default behavior */ : virtualDynamicRequirePath
-                  )})`
-                );
-                uses.commonjsHelpers = true;
+                if (shouldRemoveRequireStatement) {
+                  magicString.overwrite(node.start, node.end, `undefined`);
+                } else if (canConvertRequire) {
+                  magicString.overwrite(
+                    node.start,
+                    node.end,
+                    `${HELPERS_NAME}.commonjsRequire(${JSON.stringify(
+                      getVirtualPathForDynamicRequirePath(sourceId, commonDir)
+                    )}, ${JSON.stringify(
+                      dirname(id) === '.' ? null /* default behavior */ : virtualDynamicRequirePath
+                    )})`
+                  );
+                  uses.commonjsHelpers = true;
+                }
                 return;
               }
-              addRequireStatement(sourceId, node, scope, usesReturnValue);
+
+              if (canConvertRequire) {
+                addRequireStatement(sourceId, node, scope, usesReturnValue);
+              }
             }
 
             if (usesReturnValue) {
+              if (shouldRemoveRequireStatement) {
+                magicString.overwrite(node.start, node.end, `undefined`);
+                return;
+              }
+
               if (
                 parent.type === 'VariableDeclarator' &&
                 !scope.parent &&
@@ -205,6 +275,11 @@ export default function transformCommonjs(
               }
             } else {
               // This is a bare import, e.g. `require('foo');`
+
+              if (!canConvertRequire && !shouldRemoveRequireStatement) {
+                return;
+              }
+
               magicString.remove(parent.start, parent.end);
             }
           }
@@ -224,7 +299,18 @@ export default function transformCommonjs(
           if (!(isReference(node, parent) && !scope.contains(name))) return;
           switch (name) {
             case 'require':
-              if (isNodeRequirePropertyAccess(parent)) return;
+              if (isNodeRequirePropertyAccess(parent)) {
+                if (hasDynamicModuleForPath(id, '/', dynamicRequireModuleSet)) {
+                  if (parent.property.name === 'cache') {
+                    magicString.overwrite(node.start, node.end, `${HELPERS_NAME}.commonjsRequire`, {
+                      storeName: true
+                    });
+                    uses.commonjsHelpers = true;
+                  }
+                }
+
+                return;
+              }
 
               if (isDynamicRequireModulesEnabled && isRequireStatement(parent, scope)) {
                 magicString.appendLeft(
@@ -234,10 +320,14 @@ export default function transformCommonjs(
                   )}`
                 );
               }
+              if (isShorthandProperty(parent)) {
+                magicString.appendRight(node.end, `: ${HELPERS_NAME}.commonjsRequire`);
+              } else {
+                magicString.overwrite(node.start, node.end, `${HELPERS_NAME}.commonjsRequire`, {
+                  storeName: true
+                });
+              }
 
-              magicString.overwrite(node.start, node.end, `${HELPERS_NAME}.commonjsRequire`, {
-                storeName: true
-              });
               uses.commonjsHelpers = true;
               return;
             case 'module':
@@ -346,7 +436,8 @@ export default function transformCommonjs(
       uses.module ||
       uses.exports ||
       uses.require ||
-      uses.commonjsHelpers
+      uses.commonjsHelpers ||
+      hasRemovedRequire
     ) &&
     (ignoreGlobal || !uses.global)
   ) {