+++ packages/commonjs/test/fixtures/samples/shorthand-require/main.js
@@ -1,0 +1,7 @@
+const HOST = {
+  require
+};
+
+module.exports = {
+  HOST
+};