+++ packages/node-resolve/src/resolveImportSpecifiers.js
@@ -1,175 +1,40 @@
 import fs from 'fs';
-import path from 'path';
 import { promisify } from 'util';
+import { fileURLToPath, pathToFileURL } from 'url';
 
 import resolve from 'resolve';
 
 import { getPackageInfo, getPackageName } from './util';
 import { exists, realpath } from './fs';
 import { isDirCached, isFileCached, readCachedFile } from './cache';
+import resolvePackageExports from './package/resolvePackageExports';
+import resolvePackageImports from './package/resolvePackageImports';
+import { findPackageJson, ResolveError } from './package/utils';
 
 const resolveImportPath = promisify(resolve);
 const readFile = promisify(fs.readFile);
 
-const pathNotFoundError = (importPath, importer, subPath, pkgPath) =>
-  new Error(
-    `Could not resolve import "${importPath}" in "${importer}".` +
-      ` Package subpath "${subPath}" is not defined by "exports" in ${pkgPath}`
-  );
-
-function findExportKeyMatch(exportMap, subPath) {
-  if (subPath in exportMap) {
-    return subPath;
-  }
-
-  const matchKeys = Object.keys(exportMap)
-    .filter((key) => key.endsWith('/') || key.endsWith('*'))
-    .sort((a, b) => b.length - a.length);
-
-  for (const key of matchKeys) {
-    if (key.endsWith('*')) {
-      // star match: "./foo/*": "./foo/*.js"
-      const keyWithoutStar = key.substring(0, key.length - 1);
-      if (subPath.startsWith(keyWithoutStar)) {
-        return key;
-      }
-    }
-
-    if (key.endsWith('/') && subPath.startsWith(key)) {
-      // directory match (deprecated by node): "./foo/": "./foo/.js"
-      return key;
-    }
-
-    if (key === subPath) {
-      // literal match
-      return key;
+async function getPackageJson(importer, pkgName, resolveOptions, moduleDirectories) {
+  if (importer) {
+    const selfPackageJsonResult = await findPackageJson(importer, moduleDirectories);
+    if (selfPackageJsonResult && selfPackageJsonResult.pkgJson.name === pkgName) {
+      // the referenced package name is the current package
+      return selfPackageJsonResult;
     }
   }
-  return null;
-}
 
-function mapSubPath({ importPath, importer, pkgJsonPath, subPath, key, value }) {
-  if (typeof value === 'string') {
-    if (typeof key === 'string' && key.endsWith('*')) {
-      // star match: "./foo/*": "./foo/*.js"
-      const keyWithoutStar = key.substring(0, key.length - 1);
-      const subPathAfterKey = subPath.substring(keyWithoutStar.length);
-      return value.replace(/\*/g, subPathAfterKey);
-    }
-
-    if (value.endsWith('/')) {
-      // directory match (deprecated by node): "./foo/": "./foo/.js"
-      return `${value}${subPath.substring(key.length)}`;
-    }
-
-    // mapping is a string, for example { "./foo": "./dist/foo.js" }
-    return value;
+  try {
+    const pkgJsonPath = await resolveImportPath(`${pkgName}/package.json`, resolveOptions);
+    const pkgJson = JSON.parse(await readFile(pkgJsonPath, 'utf-8'));
+    return { pkgJsonPath, pkgJson };
+  } catch (_) {
+    return null;
   }
-
-  if (Array.isArray(value)) {
-    // mapping is an array with fallbacks, for example { "./foo": ["foo:bar", "./dist/foo.js"] }
-    return value.find((v) => v.startsWith('./'));
-  }
-
-  throw pathNotFoundError(importPath, importer, subPath, pkgJsonPath);
-}
-
-function findEntrypoint({
-  importPath,
-  importer,
-  pkgJsonPath,
-  subPath,
-  exportMap,
-  conditions,
-  key
-}) {
-  if (typeof exportMap !== 'object') {
-    return mapSubPath({ importPath, importer, pkgJsonPath, subPath, key, value: exportMap });
-  }
-
-  // iterate conditions recursively, find the first that matches all conditions
-  for (const [condition, subExportMap] of Object.entries(exportMap)) {
-    if (conditions.includes(condition)) {
-      const mappedSubPath = findEntrypoint({
-        importPath,
-        importer,
-        pkgJsonPath,
-        subPath,
-        exportMap: subExportMap,
-        conditions,
-        key
-      });
-      if (mappedSubPath) {
-        return mappedSubPath;
-      }
-    }
-  }
-  throw pathNotFoundError(importer, subPath, pkgJsonPath);
-}
-
-export function findEntrypointTopLevel({
-  importPath,
-  importer,
-  pkgJsonPath,
-  subPath,
-  exportMap,
-  conditions
-}) {
-  if (typeof exportMap !== 'object') {
-    // the export map shorthand, for example { exports: "./index.js" }
-    if (subPath !== '.') {
-      // shorthand only supports a main entrypoint
-      throw pathNotFoundError(importPath, importer, subPath, pkgJsonPath);
-    }
-    return mapSubPath({ importPath, importer, pkgJsonPath, subPath, key: null, value: exportMap });
-  }
-
-  // export map is an object, the top level can be either conditions or sub path mappings
-  const keys = Object.keys(exportMap);
-  const isConditions = keys.every((k) => !k.startsWith('.'));
-  const isMappings = keys.every((k) => k.startsWith('.'));
-
-  if (!isConditions && !isMappings) {
-    throw new Error(
-      `Invalid package config ${pkgJsonPath}, "exports" cannot contain some keys starting with '.'` +
-        ' and some not. The exports object must either be an object of package subpath keys or an object of main entry' +
-        ' condition name keys only.'
-    );
-  }
-
-  let key = null;
-  let exportMapForSubPath;
-
-  if (isConditions) {
-    // top level is conditions, for example { "import": ..., "require": ..., "module": ... }
-    if (subPath !== '.') {
-      // package with top level conditions means it only supports a main entrypoint
-      throw pathNotFoundError(importPath, importer, subPath, pkgJsonPath);
-    }
-    exportMapForSubPath = exportMap;
-  } else {
-    // top level is sub path mappings, for example { ".": ..., "./foo": ..., "./bar": ... }
-    key = findExportKeyMatch(exportMap, subPath);
-    if (!key) {
-      throw pathNotFoundError(importPath, importer, subPath, pkgJsonPath);
-    }
-    exportMapForSubPath = exportMap[key];
-  }
-
-  return findEntrypoint({
-    importPath,
-    importer,
-    pkgJsonPath,
-    subPath,
-    exportMap: exportMapForSubPath,
-    conditions,
-    key
-  });
 }
 
 async function resolveId({
   importer,
-  importPath,
+  importSpecifier,
   exportConditions,
   warn,
   packageInfoCache,
@@ -219,41 +84,70 @@ async function resolveId({
 
   let location;
 
-  const pkgName = getPackageName(importPath);
-  if (pkgName) {
-    let pkgJsonPath;
-    let pkgJson;
-    try {
-      pkgJsonPath = await resolveImportPath(`${pkgName}/package.json`, resolveOptions);
-      pkgJson = JSON.parse(await readFile(pkgJsonPath, 'utf-8'));
-    } catch (_) {
-      // if there is no package.json we defer to regular resolve behavior
-    }
-
-    if (pkgJsonPath && pkgJson && pkgJson.exports) {
-      try {
-        const packageSubPath =
-          pkgName === importPath ? '.' : `.${importPath.substring(pkgName.length)}`;
-        const mappedSubPath = findEntrypointTopLevel({
-          importer,
-          importPath,
-          pkgJsonPath,
-          subPath: packageSubPath,
-          exportMap: pkgJson.exports,
-          conditions: exportConditions
+  const pkgName = getPackageName(importSpecifier);
+  if (importSpecifier.startsWith('#')) {
+    // this is a package internal import, resolve using package imports field
+    const resolveResult = await resolvePackageImports({
+      importSpecifier,
+      importer,
+      moduleDirs: moduleDirectories,
+      conditions: exportConditions,
+      resolveId(id, parent) {
+        return resolveId({
+          importSpecifier: id,
+          importer: parent,
+          exportConditions,
+          warn,
+          packageInfoCache,
+          extensions,
+          mainFields,
+          preserveSymlinks,
+          useBrowserOverrides,
+          baseDir,
+          moduleDirectories
         });
-        const pkgDir = path.dirname(pkgJsonPath);
-        location = path.join(pkgDir, mappedSubPath);
+      }
+    });
+    location = fileURLToPath(resolveResult);
+  } else if (pkgName) {
+    // it's a bare import, find the package.json and resolve using package exports if available
+    const result = await getPackageJson(importer, pkgName, resolveOptions, moduleDirectories);
+
+    if (result && result.pkgJson.exports) {
+      const { pkgJson, pkgJsonPath } = result;
+      try {
+        const subpath =
+          pkgName === importSpecifier ? '.' : `.${importSpecifier.substring(pkgName.length)}`;
+        const pkgDr = pkgJsonPath.replace('package.json', '');
+        const pkgURL = pathToFileURL(pkgDr);
+
+        const context = {
+          importer,
+          importSpecifier,
+          moduleDirs: moduleDirectories,
+          pkgURL,
+          pkgJsonPath,
+          conditions: exportConditions
+        };
+        const resolvedPackageExport = await resolvePackageExports(
+          context,
+          subpath,
+          pkgJson.exports
+        );
+        location = fileURLToPath(resolvedPackageExport);
       } catch (error) {
-        warn(error);
-        return null;
+        if (error instanceof ResolveError) {
+          return error;
+        }
+        throw error;
       }
     }
   }
 
   if (!location) {
+    // package has no imports or exports, use classic node resolve
     try {
-      location = await resolveImportPath(importPath, resolveOptions);
+      location = await resolveImportPath(importSpecifier, resolveOptions);
     } catch (error) {
       if (error.code !== 'MODULE_NOT_FOUND') {
         throw error;
@@ -279,7 +173,7 @@ async function resolveId({
 
 // Resolve module specifiers in order. Promise resolves to the first module that resolves
 // successfully, or the error that resulted from the last attempted module resolution.
-export async function resolveImportSpecifiers({
+export default async function resolveImportSpecifiers({
   importer,
   importSpecifierList,
   exportConditions,
@@ -294,11 +188,13 @@ export async function resolveImportSpecifiers({
   rootDir,
   ignoreSideEffectsForRoot
 }) {
+  let lastResolveError;
+
   for (let i = 0; i < importSpecifierList.length; i++) {
     // eslint-disable-next-line no-await-in-loop
-    const resolved = await resolveId({
+    const result = await resolveId({
       importer,
-      importPath: importSpecifierList[i],
+      importSpecifier: importSpecifierList[i],
       exportConditions,
       warn,
       packageInfoCache,
@@ -311,9 +207,17 @@ export async function resolveImportSpecifiers({
       rootDir,
       ignoreSideEffectsForRoot
     });
-    if (resolved) {
-      return resolved;
+
+    if (result instanceof ResolveError) {
+      lastResolveError = result;
+    } else if (result) {
+      return result;
     }
+  }
+
+  if (lastResolveError) {
+    // only log the last failed resolve error
+    warn(lastResolveError);
   }
   return null;
 }