+++ packages/node-resolve/test/fixtures/exports-ts-fallback.ts
@@ -1,0 +1,3 @@
+import a from 'exports-ts-fallback/a.js';
+
+export default { a };