+++ packages/node-resolve/test/fixtures/self-package-import.js
@@ -1,0 +1,3 @@
+import selfPkgImport from 'self-package-import/a';
+
+export default selfPkgImport;