+++ packages/html/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @rollup/plugin-html ChangeLog
 
+## v0.2.2
+
+_2021-01-29_
+
+### Updates
+
+- chore: run build before publish (ce65c12)
+- chore: upgrade TypeScript (#713)
+- refactor: refactor to typescript (#634)
+- chore: update dependencies (6c8614c)
+
+## v0.2.1
+
+_2021-01-29_
+
+### Updates
+
+- chore: run build before publish (ce65c12)
+- chore: upgrade TypeScript (#713)
+- refactor: refactor to typescript (#634)
+- chore: update dependencies (6c8614c)
+
 ## v0.2.0
 
 _2020-05-02_