+++ packages/typescript/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @rollup/plugin-typescript ChangeLog
 
+## v8.1.1
+
+_2021-01-29_
+
+### Bugfixes
+
+- fix: fix plugin type declarations (#647)
+- fix: only emit tsbuildinfo file when there is something to emit (#771)
+
+## v8.1.0
+
+_2020-12-14_
+
+### Features
+
+- feat: support multiple output targets with declarations (#687)
+
+### Updates
+
+- chore: fix TypeScript warnings (#673)
+- test: code in src sub-directory (#682)
+- chore: use TypeScript 4 (#674)
+
 ## v8.0.0
 
 _2020-11-30_